requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
langchain==0.2.1
langchain-core==0.2.2
langchain-groq==0.1.3
neo4j==5.20.0
python-dotenv==1.0.1